use frontend parameters to train
