+++ ide/views.py
@@ -8,7 +8,10 @@ from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 from django.contrib.auth.models import User
 from utils.shapes import get_shapes, get_layer_shape, handle_concat_layer
-import Train
+from scripts.train import vision
+
+
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 def index(request):
     return render(request, 'index.html')
@@ -258,10 +261,11 @@ def fetch_model_history(request):
 
 @csrf_exempt
 def upload_training_data(request):
-    #print('In function upload_training_data.\n')
+    print('In function upload_training_data.\n')
     if request.method == 'POST':
         try:
             home_path = os.environ['HOME']
+
             # save current data file name in dataIndex.txt
             index_file_path = home_path + '/.VisualNN/data/dataIndex.txt'
             uploaded_file = request.FILES['file']
@@ -293,36 +297,20 @@ def start_training(request):
     if request.method == 'GET':
         try:
             print("start training...")
-            home_path = os.environ['HOME']
-            # get current model path
-            index_file_path = home_path + "/VisualNN/media/randomIndex.txt"
+            batch_size = int(request.GET['batch_size'])
+            epoch_times = int(request.GET['epoch_times'])
+            lr = float(request.GET['lr'])
+
+            index_file_path = BASE_DIR + '/test_data/randomIndex.txt'
             f = open(index_file_path, 'r')
             cur_model_name = f.readlines()[-1][:-1] + '.json'
-            model_path = home_path + "/VisualNN/media/" + cur_model_name
-            print("Current model file path: %s"%(model_path))
 
-            # get current data file path
-            data_index_file_path = home_path + "/.VisualNN/data/dataIndex.txt"
-            ff = open(data_index_file_path, 'r')
-            cur_data_file_name = ff.readlines()[-1][:-1]
-            data_path = home_path + "/.VisualNN/data/" + cur_data_file_name
-            print("Current data file path: %s"%(data_path))
+            model_path = BASE_DIR + '/test_model/' + cur_model_name	# TODO:set spcifically for each user account
+            # print("Current model file path: %s"%(model_path))
 
-            # get result file path
-            result_path = home_path + "/.VisualNN/result"
-            folder = os.path.exists(result_path)
-            if not folder:
-                os.makedirs(result_path)
-            result_file_path = result_path + "/" + cur_data_file_name.split(".")[0] + ".h5"
-            print("Current result file path: %s"%(result_file_path))
-
-            #command = "python run.py"
-            #command = "python train.py ~/.VisualNN/model/mnist.json ~/.VisualNN/data/mnist.npz result.h5"
-            #os.system(command)
-            
-            # call training function 
-            Train.trainModel(model_path, data_path, result_file_path)
-
+            data_path = BASE_DIR + '/test_data/mnist.npz'		# TODO: use uploaded data and choose by user
+            result_path = BASE_DIR + '/test_data/result'
+            vision.train_model(model_path, data_path, result_path, batch_size, epoch_times, lr)
             return JsonResponse({
                 'result': 'success'
             })
