+++ test_model/20191116124756ynhvu.json
@@ -1,0 +1,209 @@
+{
+    "class_name": "Model", 
+    "keras_version": "2.0.8", 
+    "config": {
+        "layers": [
+            {
+                "class_name": "InputLayer", 
+                "inbound_nodes": [], 
+                "config": {
+                    "dtype": "float32", 
+                    "batch_input_shape": [
+                        null, 
+                        784
+                    ], 
+                    "name": "input_1", 
+                    "sparse": false
+                }, 
+                "name": "input_1"
+            }, 
+            {
+                "class_name": "Dense", 
+                "inbound_nodes": [
+                    [
+                        [
+                            "input_1", 
+                            0, 
+                            0, 
+                            {}
+                        ]
+                    ]
+                ], 
+                "config": {
+                    "kernel_initializer": {
+                        "class_name": "Constant", 
+                        "config": {
+                            "value": 0
+                        }
+                    }, 
+                    "name": "dense_1", 
+                    "kernel_constraint": null, 
+                    "bias_regularizer": null, 
+                    "bias_constraint": null, 
+                    "activation": "linear", 
+                    "trainable": true, 
+                    "kernel_regularizer": null, 
+                    "bias_initializer": {
+                        "class_name": "Constant", 
+                        "config": {
+                            "value": 0
+                        }
+                    }, 
+                    "units": 512, 
+                    "use_bias": true, 
+                    "activity_regularizer": null
+                }, 
+                "name": "dense_1"
+            }, 
+            {
+                "class_name": "Dropout", 
+                "inbound_nodes": [
+                    [
+                        [
+                            "dense_1", 
+                            0, 
+                            0, 
+                            {}
+                        ]
+                    ]
+                ], 
+                "config": {
+                    "rate": 0.5, 
+                    "trainable": true, 
+                    "name": "dropout_1"
+                }, 
+                "name": "dropout_1"
+            }, 
+            {
+                "class_name": "Dense", 
+                "inbound_nodes": [
+                    [
+                        [
+                            "dropout_1", 
+                            0, 
+                            0, 
+                            {}
+                        ]
+                    ]
+                ], 
+                "config": {
+                    "kernel_initializer": {
+                        "class_name": "Constant", 
+                        "config": {
+                            "value": 0
+                        }
+                    }, 
+                    "name": "dense_2", 
+                    "kernel_constraint": null, 
+                    "bias_regularizer": null, 
+                    "bias_constraint": null, 
+                    "activation": "linear", 
+                    "trainable": true, 
+                    "kernel_regularizer": null, 
+                    "bias_initializer": {
+                        "class_name": "Constant", 
+                        "config": {
+                            "value": 0
+                        }
+                    }, 
+                    "units": 10, 
+                    "use_bias": true, 
+                    "activity_regularizer": null
+                }, 
+                "name": "dense_2"
+            }, 
+            {
+                "class_name": "Dropout", 
+                "inbound_nodes": [
+                    [
+                        [
+                            "dense_2", 
+                            0, 
+                            0, 
+                            {}
+                        ]
+                    ]
+                ], 
+                "config": {
+                    "rate": 0.5, 
+                    "trainable": true, 
+                    "name": "dropout_2"
+                }, 
+                "name": "dropout_2"
+            }, 
+            {
+                "class_name": "Dense", 
+                "inbound_nodes": [
+                    [
+                        [
+                            "dropout_2", 
+                            0, 
+                            0, 
+                            {}
+                        ]
+                    ]
+                ], 
+                "config": {
+                    "kernel_initializer": {
+                        "class_name": "Constant", 
+                        "config": {
+                            "value": 0
+                        }
+                    }, 
+                    "name": "dense_3", 
+                    "kernel_constraint": null, 
+                    "bias_regularizer": null, 
+                    "bias_constraint": null, 
+                    "activation": "linear", 
+                    "trainable": true, 
+                    "kernel_regularizer": null, 
+                    "bias_initializer": {
+                        "class_name": "Constant", 
+                        "config": {
+                            "value": 0
+                        }
+                    }, 
+                    "units": 10, 
+                    "use_bias": true, 
+                    "activity_regularizer": null
+                }, 
+                "name": "dense_3"
+            }, 
+            {
+                "class_name": "Activation", 
+                "inbound_nodes": [
+                    [
+                        [
+                            "dense_3", 
+                            0, 
+                            0, 
+                            {}
+                        ]
+                    ]
+                ], 
+                "config": {
+                    "activation": "softmax", 
+                    "trainable": true, 
+                    "name": "activation_1"
+                }, 
+                "name": "activation_1"
+            }
+        ], 
+        "input_layers": [
+            [
+                "input_1", 
+                0, 
+                0
+            ]
+        ], 
+        "name": "Untitled", 
+        "output_layers": [
+            [
+                "activation_1", 
+                0, 
+                0
+            ]
+        ]
+    }, 
+    "backend": "tensorflow"
+}
